+++ lib/src/video_player_initializer/file_initializer.dart
@@ -1,17 +1,18 @@
 import 'package:flutter/cupertino.dart';
-import 'package:omni_video_player/src/api/hls_video_api.dart';
+import 'package:flutter/foundation.dart';
 import 'package:omni_video_player/src/controllers/default_playback_controller.dart';
+import 'package:omni_video_player/src/utils/logger.dart';
 import 'package:omni_video_player/src/video_player_initializer/video_player_initializer_factory.dart';
 import 'package:omni_video_player/omni_video_player.dart';
 import 'package:video_player/video_player.dart' show VideoPlayer;
 
-class NetworkInitializer implements IVideoPlayerInitializerStrategy {
+class FileInitializer implements IVideoPlayerInitializerStrategy {
   final VideoPlayerConfiguration options;
   final VideoPlayerCallbacks callbacks;
   final GlobalPlaybackController? globalController;
   final void Function()? onErrorCallback;
 
-  NetworkInitializer({
+  FileInitializer({
     required this.options,
     required this.callbacks,
     this.globalController,
@@ -20,28 +21,17 @@ class NetworkInitializer implements IVideoPlayerInitializerStrategy {
 
   @override
   Future<OmniPlaybackController?> initialize() async {
+    if (kIsWeb) {
+      logger.w(
+          "File initializer is not supported on web, see DOC: https://pub.dev/packages/video_player_web");
+      return null;
+    }
+
     try {
-      final isHlsVideo = await HlsVideoApi.isHlsUri(
-          options.videoSourceConfiguration.videoUrl!);
-
-      Map<OmniVideoQuality, Uri>? qualitiesMap;
-      MapEntry<OmniVideoQuality, Uri>? currentQualityEntry;
-
-      if (isHlsVideo) {
-        qualitiesMap = await HlsVideoApi.extractHlsVariantsByQuality(
-            options.videoSourceConfiguration.videoUrl!,
-            options.videoSourceConfiguration.availableQualities);
-        currentQualityEntry = HlsVideoApi.selectBestQualityVariant(qualitiesMap,
-            preferredQualities:
-                options.videoSourceConfiguration.preferredQualities);
-      }
-
       final controller = await DefaultPlaybackController.create(
-        videoUrl: (currentQualityEntry != null)
-            ? currentQualityEntry.value
-            : options.videoSourceConfiguration.videoUrl!,
+        videoUrl: null,
         dataSource: null,
-        file: null,
+        file: options.videoSourceConfiguration.videoFile,
         audioUrl: null,
         isLive: false,
         globalController: globalController,
@@ -52,9 +42,6 @@ class NetworkInitializer implements IVideoPlayerInitializerStrategy {
         callbacks: callbacks,
         type: options.videoSourceConfiguration.videoSourceType,
         globalKeyPlayer: options.globalKeyPlayer,
-        qualityUrls: qualitiesMap,
-        currentVideoQuality:
-            (currentQualityEntry != null) ? currentQualityEntry.key : null,
       );
 
       controller.sharedPlayerNotifier.value = Hero(
@@ -67,7 +54,8 @@ class NetworkInitializer implements IVideoPlayerInitializerStrategy {
 
       callbacks.onControllerCreated?.call(controller);
       return controller;
-    } catch (e) {
+    } catch (e, st) {
+      logger.e(e, stackTrace: st);
       onErrorCallback?.call();
       return null;
     }