+++ lib/src/video_player_initializer/video_player_initializer.dart
@@ -194,6 +194,9 @@ class VideoPlayerInitializerState extends State<VideoPlayerInitializer>
 
       case VideoSourceType.asset:
         return widget.options.customPlayerWidgets.thumbnail;
+
+      case VideoSourceType.file:
+        return widget.options.customPlayerWidgets.thumbnail;
     }
   }
 }